+++ R/retrocompatibility.R
@@ -1,4 +1,4 @@
-#' Update labelled data to version 2.0.0
+#' Update labelled data to last version
 #'
 #' Labelled data imported with \code{haven} version 1.1.2 or before or
 #' created with \code{labelled} version 1.1.0 or before was using
@@ -7,8 +7,12 @@
 #' Since version 2.0.0 of these two packages, "haven_labelled" and
 #' "haven_labelled_spss" are used instead.
 #'
+#' Since haven 2.3.0, "haven_labelled" class has been evolving
+#' using now \pkg{vctrs} package.
+#'
 #' \code{update_labelled} convert labelled vectors
-#' from the old to the new classes.
+#' from the old to the new classes and to reconstruct all
+#' labelled vectors with the last version of the package.
 #'
 #' @param x An object (vector or data.frame) to convert.
 #' @export
@@ -37,6 +41,24 @@ update_labelled.labelled <- function(x) {
     }
   }
   x
+}
+
+#' @rdname update_labelled
+#' @export
+update_labelled.haven_labelled_spss <- function(x) {
+  labelled_spss(
+    x, labels = val_labels(x), label = var_label(x),
+    na_values = na_values(x), na_range = na_range(x)
+  )
+}
+
+
+#' @rdname update_labelled
+#' @export
+update_labelled.haven_labelled <- function(x) {
+  labelled(
+    x, labels = val_labels(x), label = var_label(x)
+  )
 }
 
 #' @rdname update_labelled